requirements.txt
dnspython == 2.*
requests == 2.*
tld == 0.12.*
jwcrypto == 1.*
cryptography == 36.*